refactor(step-5): move bear-urls.ts and its test to src/infra/
bear-urls.ts joins database.ts in the infra/ layer. Both are the only
modules that speak to the outside world (Bear's SQLite DB and the
x-callback-url URL scheme via 'open'). Their own imports go up one
level (../config.js, ../logging.js); the co-located test imports
remain sibling-relative (./bear-urls.js). Callers (main.ts, utils.ts)
now import from ./infra/bear-urls.js.

Note: the utils <-> bear-urls cycle is still present at this step
because utils.ts still holds handleNoteTextUpdate, which calls
buildBearUrl and executeBearXCallbackApi. That cycle dissolves in
step 9 when handleNoteTextUpdate moves into tools/note-tools.ts and
utils.ts is deleted.
